refactor: rename modal-screen-options to create-options and update imports
- Rename modal-screen-options.tsx to create-options.tsx for clarity
- Update import path in _layout.tsx
- Fix ScreenWrapper import path in modal-wrapper.tsx
- Change modal wrapper safe area from 'top' to 'both'
